+++ src/main/java/fr/eni/PizzaOnline/dal/CheeseDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import fr.eni.PizzaOnline.bo.Cheese;
 
-public interface CheeseDAO extends CrudRepository<Cheese, Long>{
+public interface CheeseDAO extends CrudRepository<Cheese, Integer>{
 
 }
