+++ src/main/java/fr/eni/PizzaOnline/dal/IngredientDAO.java
@@ -3,6 +3,6 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.Ingredient;
 
-public interface IngredientDAO extends CrudRepository<Ingredient, Long>{
+public interface IngredientDAO extends CrudRepository<Ingredient, Integer>{
 
 }
