+++ src/main/java/fr/eni/PizzaOnline/dal/ReductionCodeDAO.java
@@ -3,6 +3,6 @@ package fr.eni.PizzaOnline.dal;
 import org.springframework.data.repository.CrudRepository;
 import fr.eni.PizzaOnline.bo.ReductionCode;
 
-public interface ReductionCodeDAO extends CrudRepository<ReductionCode, Long>{
+public interface ReductionCodeDAO extends CrudRepository<ReductionCode, Integer>{
 
 }
